Added master for new Post (wip)
