+++ app/src/App.js
@@ -13,7 +13,7 @@ import Roadmap from './pages/Roadmap'
 import Signals from './pages/Signals'
 import Account from './pages/Account'
 import EventVotes from './pages/EventVotes'
-import EventVotesNew from './pages/EventVotesNew'
+import EventVotesNew from './pages/EventVotesNew/EventVotesNew'
 import SideMenu from './components/SideMenu'
 import MobileMenu from './components/MobileMenu'
 import withTracker from './withTracker'
