+++ app/src/pages/EventVotesNew/EventVotesNew.css
